backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9  # postgresql adapter
python-dotenv==1.0.0
requests==2.31.0
gunicorn==21.2.0  # production server
whitenoise==6.6.0  # static files in production 